WIP: adding filter options in search bar
